+++ thetawave_lib/src/lib.rs
@@ -3,6 +3,7 @@ pub mod components;
 pub mod constants;
 pub mod entities;
 pub mod events;
+pub mod motion;
 pub mod resources;
 pub mod spawn;
 pub mod spawnable;
