+++ thetawave_lib/src/motion/components/mod.rs
@@ -1,0 +1,4 @@
+mod hitbox;
+mod motion;
+
+pub use self::{hitbox::Hitbox2DComponent, motion::Motion2DComponent};
