+++ thetawave_lib/src/motion/mod.rs
@@ -1,0 +1,4 @@
+//! `thetawave_lib` motion module
+
+pub mod components;
+pub mod systems;
